requirements.txt
streamlit
langchain
langchain-google-genai
langchain-openai
langgraph
python-dotenv
requests
yfinance
pandas
faker
plotly
markdown
beautifulsoup4
